==> 4_figures_paper.rmd <==
---
title: "Contextual modulations in low- and high-levels of visual processing hierarchy"
date: "Jan 22nd 2022"
output:
  rmdformats::downcute:
    self_contained: true
    thumbnails: true
    lightbox: true
    gallery: false
    highlight: tango
    code_folding: hide
editor_options: 
  chunk_output_type: console
---


```{r setup, include = FALSE, eval=TRUE, cache = FALSE}
# Loading packages, defining some default options for code chunks, 
# and a custom ggplot theme for the graphs.

#library(rstan)
library(tidyverse)
library(tidybayes)
library(magrittr, warn.conflicts = FALSE)
library(papaja)
library(knitr)
library(brms, warn.conflicts = FALSE)
library(boot)
library(readxl)
library(Hmisc, warn.conflicts = FALSE)
library(modelr)
library(reshape2, warn.conflicts = FALSE)
#library(rethinking, warn.conflicts = FALSE)
library(broom, warn.conflicts = FALSE)
library(glue)
library(BayesFactor)
library(bayestestR)
library(patchwork)
# ggplot theme 
muctheme <- theme_classic() +
  theme(
    axis.text=element_text(size=20, face="bold"),
    axis.title=element_text(size=25, face="bold"),
    strip.text= element_text(size=10),
    legend.title=element_blank(),
    #panel.background = element_rect(fill = 'gray96'),
    #panel.border = element_rect(size = 2),
    strip.background = element_rect(fill = 'white', color = "white"),
    #strip.placement = 'outside',
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank())

knitr::opts_chunk$set(echo = TRUE,
                      eval = TRUE,
                      message = FALSE,
                      warning = FALSE)
```


# Participant Data

```{r participant data}
load("subinfo_active_123rd.rda")
load("lo_clean123.rda")
```


```{r some functions}
subj_count <- function(x) x %>% group_by(subject) %>% tally() %>% summarise(n=n())
```

## Practice Fail/Pass
```{r Practice data, eval=FALSE}
practable <- subinfo123rd %>% group_by(Practice) %>% tally()
practable
subinfo_active_123rd <- subinfo123rd %>% filter(Practice==1)
save(subinfo_active_123rd, file="subinfo_active_123rd.rda")
```

## Sex Distribution
```{r sex dist, eval=FALSE}

sextable <- lo_clean123 %>% group_by(subject,gender) %>% tally()

sextable %>% group_by(gender) %>% tally()

subinfo_active_123rd$gender <- sextable$gender

```

## Handedness

```{r hand, eval=FALSE}
handedness <- subinfo_active_123rd %>% mutate(hand = case_when(EHI > 0 ~ "righty",
                                    EHI < 0 ~ "lefty")) %>% group_by(hand) %>% tally()
handedness
```

## Age Distribution
```{r Age, eval=FALSE}
load("lo_clean_123rd.rda")
age <- lo_clean123 %>% group_by(subject,age) %>% tally()
age2 <- dfl %>% group_by(age) %>% tally() 


subinfo_active_123rd$age <- age$age

age %>% ungroup %>%  summarise(mean=mean(age), sd=sd(age))

age %>% ggplot(aes(x = age)) +
  geom_histogram(binwidth=.5, color="black", fill="grey")+
  muctheme+
  geom_vline(aes(xintercept=mean(age)), size=2, color="brown",show.legend=FALSE)
```

Mean age = `r round(mean(subinfo$age),2)`
SD = `r round(sd(subinfo$age),2)`


## Visual Acuity (FrACT)
```{r VA, eval=FALSE}
subinfo_active_123rd%>% ggplot(aes(x = `VA(dec)`)) +
  geom_histogram(binwidth=.05, color="black", fill="grey")+
  muctheme+
  geom_vline(aes(xintercept=mean(`VA(dec)`)), size=2, color="brown",show.legend=FALSE)

meanVA=round(mean(subinfo_active_123rd$`VA(dec)`),2) 
sdVA=round(sd(subinfo_active_123rd$`VA(dec)`),2)
```

Mean VA(dec) = `r meanVA`
SD = `r sdVA`

## BFRT (Benton Face Recognition Task)
```{r Benton, eval=FALSE}
subinfo_active_123rd %>% ggplot(aes(x = Benton)) +
  geom_histogram(binwidth=.5, color="black", fill="grey")+
  muctheme+
  geom_vline(aes(xintercept=mean(Benton)), size=2, color="brown",show.legend=FALSE)+
  xlab("BFRT Score")
```

Mean = `r mean(subinfo_active_123rd$Benton, na.rm=T)`
SD = `r sd(subinfo_active_123rd$Benton, na.rm=T)`

```{r big model fit, eval=FALSE}
load("d_lohi1234.rda")

df <-  d_lohi1234

df_trns <- df %>% mutate(iv2= case_when(task=="low"~log10(iv),
                                        TRUE~iv)) %>%
  group_by(task) %>%
  mutate(iv3=(iv2-mean(iv2))/sd(iv2),
         task=as.factor(task),
         guess=case_when(task=="low"~1,
                         task!="low"~0))

fit_invlogit_1234th <- brm(
  bf(dv ~ 0.5*guess + (1-0.5*guess)*inv_logit(eta),
     eta ~ iv3*task*cond + (iv3*task*cond||subject), nl=TRUE),
  data=df_trns,
  family=bernoulli("identity"),
  prior=c(prior(normal(0,5), nlpar="eta"),
          prior(lognormal(0,.1), nlpar="eta", class="sd", group="subject")),
  cores=12)
## This needs to be updated with correct # of iterations, etc..
save(fit_invlogit_1234th, file="fit_invlogit_1234th.rda", compress="xz")
```


# Psychometric Functions

## Example subject psych fitting

```{r example subject}

load("brmslsETA_1234th.rda") # These are the slope/intercept values I obtained using the inv_logit calculation. Now we check if they match with sumcoefs
load("preds_invlogit_1234th.rda")
load("d_lohi1234.rda")

etas2 <- brmseta_1234 %>% mutate(task=case_when(task=="inv"~"Inverted Face",
                                           task=="up"~"Upright Face",
                                           task=="low"~"Contrast Detection"),
                         cond = case_when(cond=="iso"~"Isolated",
                                               cond=="diff"~"Different",
                                               cond=="same"~"Same"))
etaslow <- etas2 %>% filter(task=="Contrast Detection")
etasup <- etas2 %>% filter(task=="Upright Face")
etasinv <- etas2 %>% filter(task=="Inverted Face")




etasLH01 <- etas2 %>% filter(subject=="LH01")



df_trns <- d_lohi1234 %>% mutate(iv2= case_when(task=="low"~log10(iv),
                                        TRUE~iv)) %>%
  group_by(task) %>%
  mutate(iv3=(iv2-mean(iv2))/sd(iv2),
         task=as.factor(task),
         guess=case_when(task=="low"~1,
                         task!="low"~0))

d_av<- df_trns %>% group_by(iv3, cond, task, subject) %>% 
  summarise(mean_p = mean(dv))

d_count <- df_trns %>% group_by(iv3, cond, task, subject) %>% 
  summarise(n=n(), mean=mean(dv))

tasks3n <- d_count %>% filter(subject=="LH01")

tasks3av<- d_av %>% filter(subject=="LH01")
tasks3ls <- etas2 %>% filter(subject=="LH01")
pred_df <- pred_df %>% mutate(hline = case_when(task=="low"~.75,
                                                task!="low"~.5))

pred_df <- pred_df %>% mutate(task = case_when(task=="low"~"Contrast Detection",
                                              task=="upright"~"Upright Face",
                                              task=="inverted"~"Inverted Face"),
                              cond = case_when(cond=="iso"~"Isolated",
                                               cond=="diff"~"Different",
                                               cond=="same"~"Same"))
tasks3n <- tasks3n %>% mutate(task = case_when(task=="low"~"Contrast Detection",
                                              task=="upright"~"Upright Face",
                                              task=="inverted"~"Inverted Face"),
                             cond = case_when(cond=="iso"~"Isolated",
                                              cond=="diff"~"Different",
                                              cond=="same"~"Same"))

pred_df$task <- ordered(pred_df$task, levels = c("Upright Face", "Inverted Face","Contrast Detection"))
pred_df$cond <- ordered(pred_df$cond, levels = c("Different","Same","Isolated"))
tasks3ls$task <- ordered(tasks3ls$task, levels = c("Upright Face", "Inverted Face","Contrast Detection"))
tasks3n$task <- ordered(tasks3n$task, levels = c("Upright Face", "Inverted Face","Contrast Detection"))


pred_df %>% filter(subject=="LH01") %>% ggplot(aes(x = iv3, y = Estimate, color=cond)) + 
  geom_line() + 
  geom_ribbon(aes(ymin = Q2.5, ymax = Q97.5, fill=cond), alpha = .4, color="NA") + 
  geom_count(data = tasks3n, aes(x = iv3, y = mean,size=n), show.legend=FALSE, alpha=.8) + 
  scale_size_area(max_size=5)+
  ggthemes::scale_fill_colorblind() + 
  ggthemes::scale_color_colorblind() +
  #geom_hline(yintercept=.5)+
  geom_hline(aes(yintercept=hline))+
  geom_vline(data=tasks3ls, aes(xintercept=location, color=cond))+
  #geom_vline(data=tasks3ls, aes(xintercept=location+slope/2, color=cond), linetype="dashed")+
  #geom_vline(data=tasks3ls, aes(xintercept=location-slope/2, color=cond), linetype="dashed")+
  facet_wrap(~task)+
  #ggtitle("LH01")+
  theme_classic()+
  xlab("                   Dissimilarity (z)                          Contrast (log & z)")+
  ylab(expression(atop(bold("Dependent Variable"),atop(italic("(Proportion 'Different' or Accuracy)")))))+
  #ylab("Dependent Variable \n (Proportion 'Different' or Accuracy)")+
  muctheme+
  theme(axis.text.x = element_text(size=13),
        axis.text.y = element_text(size=13),
        strip.text = element_text(size=20, face="bold"),
        legend.title=element_blank(),
        legend.text=element_text(size=20, face="bold"))+
  xlim(-3,3)+
  theme(axis.title.x=element_text(size=20, face="bold", hjust=0), axis.title=element_text(size=16, face="bold"))

```

```{r, preds for all subjects, eval=FALSE}

pred_df %>% filter(task=="Upright Face") %>% ggplot(aes(x=iv3, y=Estimate, color=cond))+
  geom_line()+
  geom_ribbon(aes(ymin=Q2.5, ymax=Q97.5, fill=cond), alpha=.1, color="NA")+
  ggthemes::scale_fill_colorblind() + 
  ggthemes::scale_color_colorblind() + 
  facet_wrap(~subject, scales="free")+
  geom_vline(data=etasup, aes(xintercept=location, color=cond))+
  geom_hline(yintercept=.5)+
#  geom_vline(data=etasup, aes(xintercept=location+scale/2, color=cond), linetype="dashed")+
#  geom_vline(data=etasup, aes(xintercept=location-scale/2, color=cond), linetype="dashed")+
  theme_classic()+
  ggtitle("Upright Face Psychometric Functions")+
  theme(legend.position="none")+
  #theme(strip.text.x = element_text(size = 8, colour = "orange", angle = 90))
  theme(strip.text.x = element_blank(),
        strip.background = element_blank())+
  geom_text(data=etasup,x=-Inf,y=+Inf,aes(label=subject),
            vjust = "inward", hjust = "inward", colour="black", size=2)+
  theme(axis.text.x = element_text(size=5),
        axis.text.y = element_text(size=5),
        strip.text = element_text(size=1))

pred_df %>% filter(task=="Inverted Face") %>%  ggplot(aes(x=iv3, y=Estimate, color=cond))+
  geom_line()+
  geom_ribbon(aes(ymin=Q2.5, ymax=Q97.5, fill=cond), alpha=.1, color="NA")+
  ggthemes::scale_fill_colorblind() + 
  ggthemes::scale_color_colorblind() + 
  facet_wrap(~subject, scales="free")+
  geom_vline(data=etasinv, aes(xintercept=location, color=cond))+
  geom_hline(yintercept=.5)+
#  geom_vline(data=etasinv, aes(xintercept=location+scale/2, color=cond), linetype="dashed")+
#  geom_vline(data=etasinv, aes(xintercept=location-scale/2, color=cond), linetype="dashed")+
  theme_classic()+
  ggtitle("Inverted Face Psychometric Functions")+
  theme(legend.position="none")+
  #theme(strip.text.x = element_text(size = 8, colour = "orange", angle = 90))
  theme(strip.text.x = element_blank(),
        strip.background = element_blank())+
  geom_text(data=etasinv,x=-Inf,y=+Inf,aes(label=subject),
            vjust = "inward", hjust = "inward", colour="black", size=2)+
  theme(axis.text.x = element_text(size=5),
        axis.text.y = element_text(size=5),
        strip.text = element_text(size=1))

pred_df %>% filter(task=="Contrast Detection") %>%  ggplot(aes(x=iv3, y=Estimate, color=cond))+
  geom_line()+
  geom_ribbon(aes(ymin=Q2.5, ymax=Q97.5, fill=cond), alpha=.1, color="NA")+
  ggthemes::scale_fill_colorblind() + 
  ggthemes::scale_color_colorblind() + 
  facet_wrap(~subject, scales="free")+
  geom_vline(data=etaslow, aes(xintercept=location, color=cond))+
  geom_hline(yintercept=.5)+
#  geom_vline(data=etasinv, aes(xintercept=location+scale/2, color=cond), linetype="dashed")+
#  geom_vline(data=etasinv, aes(xintercept=location-scale/2, color=cond), linetype="dashed")+
  theme_classic()+
  ggtitle("Contrast Detection Psychometric Functions")+
  theme(legend.position="none")+
  #theme(strip.text.x = element_text(size = 8, colour = "orange", angle = 90))
  theme(strip.text.x = element_blank(),
        strip.background = element_blank())+
  geom_text(data=etasinv,x=-Inf,y=+Inf,aes(label=subject),
            vjust = "inward", hjust = "inward", colour="black", size=2)+
  theme(axis.text.x = element_text(size=5),
        axis.text.y = element_text(size=5),
        strip.text = element_text(size=1))



# These plots show that the location is exactly at .5 or .75 point, which is a great thing! Because when
# calculated using inv_logit formula, sometimes they were slightly off.
```

## Summarized psych fitting

```{r}
load("preds_1234th_general.rda")
load("summary_etas_1234th.rda")


summary_etas <- summary_etas_1234th %>% mutate(task = case_when(task=="low"~"Grating",
                                                 task=="up"~"Upright Face",
                                                 task=="inv"~"Inverted Face"),
                                cond = case_when(cond=="iso"~"Isolated",
                                                 cond=="diff"~"Different",
                                                 cond=="same"~"Same"))
preds_general <- preds_1234th_general %>% mutate(task = case_when(task=="low" ~ "Grating",
                                                           task=="upright" ~ "Upright Face",
                                                           task=="inverted" ~ "Inverted Face"),
                                          cond = case_when(cond=="iso"~"Isolated",
                                                           cond=="diff"~"Different",
                                                           cond=="same"~"Same"))

preds_general$task <- ordered(preds_general$task, levels = c("Upright Face", "Inverted Face","Grating"))
preds_general$cond <- ordered(preds_general$cond, levels = c("Different","Same","Isolated"))
summary_etas$task <- ordered(summary_etas$task, levels = c("Upright Face", "Inverted Face","Grating"))

preds_general <- preds_general %>% mutate(hline = case_when(task=="Grating"~.75,
                                                task!="Grating"~.5))

preds_general %>%  ggplot(aes(x = iv3, y = Estimate, color=cond)) + 
  geom_line() + 
  geom_ribbon(aes(ymin = Q2.5, ymax = Q97.5, fill=cond), alpha = .4, color="NA") + 
  #geom_count(data = tasks3n, aes(x = iv3, y = mean,size=n), show.legend=FALSE, alpha=.8) + 
  scale_size_area(max_size=5)+
  ggthemes::scale_fill_colorblind() + 
  ggthemes::scale_color_colorblind() +
  #geom_hline(yintercept=.5)+
  geom_hline(aes(yintercept=hline))+
  geom_vline(data=summary_etas, aes(xintercept=location, color=cond))+
  #geom_vline(data=tasks3ls, aes(xintercept=location+slope/2, color=cond), linetype="dashed")+
  #geom_vline(data=tasks3ls, aes(xintercept=location-slope/2, color=cond), linetype="dashed")+
  facet_wrap(~task)+
  #ggtitle("LH01")+
  theme_classic()+
  xlab("             Dissimilarity (z)               Contrast (log & z)")+
  ylab(expression(atop(bold("Dependent Variable"),atop(italic("(Proportion 'Different' or Accuracy)")))))+
  
  #ylab("Dependent Variable \n (Proportion 'Different' or Accuracy)")+
  muctheme+
  theme(axis.text.x = element_text(size=13),
        axis.text.y = element_text(size=13),
        strip.text = element_text(size=20, face="bold"),
        legend.title=element_blank(),
        legend.text=element_text(size=20, face="bold"))+
  xlim(-3,3)+
  theme(axis.title.x=element_text(size=20, face="bold", hjust=0), axis.title=element_text(size=16, face="bold"))


```



# Upright-Isolated Proportion Difference (for Supplementary)

```{r fig 5 up-iso}

load("d_lohi1234.rda")

df <- d_lohi1234


df_up_iso <- df %>% filter(task == "upright",
                           cond == "iso")

mod <- glm(dv ~ iv ,binomial(),df_up_iso)
mdata <- expand.grid(iv= seq(0,100, l = 101))
mdata <- add_column(mdata, fit = predict(mod, newdata = mdata, type = 'response'))
fam <- family(mod)
ilink <- fam$linkinv
ilink <- family(mod)$linkinv
mdata <- bind_cols(mdata, setNames(as_tibble(predict(mod, mdata, se.fit = TRUE)[1:2]),
                                   c('fit_link','se_link')))
mdata <- mutate(mdata,
                fit_resp  = ilink(fit_link),
                right_upr = ilink(fit_link + (2 * se_link)),
                right_lwr = ilink(fit_link - (2 * se_link)))


dhu_avg <- df_up_iso %>% 
  group_by(iv) %>%
  summarise(mean= mean(dv),
            sd=sd(dv),
            se=sd/sqrt(length(unique(df_up_iso$subject))))


mdata  %>% ggplot(aes(x=iv, y=fit))+
  geom_line(size=1.3, show.legend=T, alpha=.3)+
  geom_point(size=5, data=dhu_avg, aes(x=iv, y=mean), alpha=.5)+
  geom_errorbar(data=dhu_avg, inherit.aes=F, aes(x=iv, ymax=mean+se, ymin=mean-se), width=2.5, alpha=.3)+
  geom_ribbon(aes(ymin=right_lwr, ymax=right_upr), alpha=.25)+
  theme_classic() + 
  ggthemes::scale_color_colorblind()+
  ggthemes:: scale_fill_colorblind()+
  scale_x_continuous(breaks=c(0,24,36,53,80))+
  scale_y_continuous(limits=c(0,1))+
  xlab("Dissimilarity (%)")+
  ylab("Proportion of 'Different' Responses")+
  theme(legend.position="none")+
  theme(axis.text=element_text(size=14, face="bold"), axis.title=element_text(size=16, face="bold"))
```

## Figure 5 - Curve fitting

```{r fig 5 curve fitting}


fit <- lm(mean ~ iv, data=dhu_avg)
fit2 <- lm(mean ~ poly(iv,2), data=dhu_avg)
fit3 <- lm(mean ~ poly(iv,3), data=dhu_avg)

xx <- seq(0,100, length=50)
plot(dhu_avg$iv,dhu_avg$mean,pch=19,ylim=c(0,1))
lines(xx, predict(fit, data.frame(iv=xx)), col="red")
lines(xx, predict(fit2, data.frame(iv=xx)), col="green")
lines(xx, predict(fit3, data.frame(iv=xx)), col="blue")

summary(fit)
summary(fit2)
summary(fit3) # we see that this is the best-fitting model.


```


# Contextual Modulations in face task (for Supplementary)

```{r figure 6, hi-task, conteff}
Path <- getwd()

load(paste(Path,"/hi_clean1234.rda",sep=""))
dfhacc <- hi_clean1234

dhdiff <- dfhacc %>% select(TrialNo, dissim, ori, subject, context, diff)

dhd <- dhdiff %>% group_by(subject, ori, dissim, context) %>%
  summarise(diff=mean(diff))


dhd2 <- dhd %>% pivot_wider(names_from="context", values_from="diff") %>%
  mutate(condeff = diff-same) %>% 
  group_by(ori, dissim) %>%
  summarise(secondeff = sd(condeff)/sqrt(length(unique(dhd$subject))),
            condeff=mean(condeff)) %>%
  mutate(ori=as.character(ori)) %>%
  mutate(Orientation = case_when(ori=="0"~ "Upright",
                                 ori=="180"~"Inverted"))


line.df <-  dhd2[dhd2$dissim > 5, ]

dhd2$Orientation <- ordered(dhd2$Orientation, levels = c("Upright","Inverted"))


dhd2 %>% ggplot(aes(x=dissim, y=condeff, group=Orientation, color=Orientation))+
  geom_point(size=4)+
  geom_line(data=line.df, size=1)+
  geom_errorbar(aes(ymin=condeff-secondeff, ymax=condeff+secondeff), width=1)+
  ylab(expression(atop(bold("Contextual Modulation Magnitude"),atop(italic(Prop.Diff[DifferentContext]-Prop.Diff[SameContext])))))+
  xlab("Dissimilarity (%)")+
  theme_classic()+
  scale_fill_manual(values=c("black","grey")) + 
  scale_color_manual(values=c("black","grey"))+
  scale_x_continuous(breaks=c(0,24,36,53,80))+
  theme(axis.text=element_text(size=20, face="bold"), axis.title=element_text(size=25, face="bold"))+
  theme(legend.title=element_text(size=20),legend.text=element_text(size=25))

```

## Modeling stuff for later
```{r fig 6 model, general model checks for BF01, eval=FALSE}

dhd3 <- dhd %>% pivot_wider(names_from="context", values_from="diff") %>%
  mutate(condeff = diff-same) %>% 
  group_by(ori, dissim, subject) %>%
  summarise(condeff=mean(condeff)) %>%
  mutate(ori=as.factor(ori)) %>%
  mutate(Orientation = case_when(ori=="0"~ "Upright",
                                 ori=="180"~"Inverted"))

fit_ce2 <- brm(
  bf(condeff ~ 0 + dissim*ori + (0 + dissim*ori||subject)),
  prior=c(prior(normal(0, 10), class = b),
          prior(cauchy(0, 10), class = sd),
          prior(cauchy(0, 10), class = sigma)),
  data=dhd3,
  sample_prior=TRUE,
  iter=1000,
  warmup=500,
  cores = parallel::detectCores())

save(fit_ce2, file="fit_ce2.rds", compress="xz")


fit_ce3 <- brm(
  bf(condeff ~ 0 + dissim*ori + (0 + dissim*ori||subject)),
  prior=c(prior(normal(0,1), class=b),
          prior(cauchy(0,1), class=sd),
          prior(cauchy(0,1), class=sigma)),
  data=dhd3,
  sample_prior=TRUE,
  iter=1000,
  warmup=500,
  cores=parallel::detectCores())


save(fit_ce3, file="fit_ce3.rds", compress="xz")

fit_ce4 <- brm(
  bf(condeff ~ 0 + dissim*ori + (0 + dissim*ori||subject)),
  prior=c(prior(normal(.2,.1), class=b),
          prior(cauchy(0,.1), class=sd),
          prior(cauchy(0,.1), class=sigma)),
  data=dhd3,
  sample_prior=TRUE,
  iter=1000,
  warmup=500,
  cores=parallel::detectCores())


save(fit_ce4, file="fit_ce4.rds", compress="xz")


fit_ce5 <- brm(
  bf(condeff ~ 0 + dissim*ori + (0 + dissim*ori||subject)),
  prior=c(prior(normal(.2,.1), class=b),
          prior(cauchy(0,.1), class=sd),
          prior(cauchy(0,.1), class=sigma)),
  data=dhd3,
  sample_prior=TRUE,
  iter=1000,
  warmup=500,
  cores=parallel::detectCores())


save(fit_ce5, file="fit_ce5.rds", compress="xz")

fit_ce6 <- brm(
  bf(condeff ~ 0 + dissim*ori + (0 + dissim*ori||subject)),
  prior=c(prior(normal(0,.005), class=b, coef="dissim"),
          prior(cauchy(0,.1), class=sd),
          prior(cauchy(0,.1), class=sigma)),
  data=dhd3,
  sample_prior=TRUE,
  iter=1000,
  warmup=500,
  cores=parallel::detectCores())

save(fit_ce6, file="fit_ce6.rds", compress="xz")


fit_ce7 <- brm(
  bf(condeff ~ 0 + dissim*ori + (0 + dissim*ori||subject)),
  prior=c(prior(normal(0,.1), class=b, coef="ori0"),
          prior(normal(0,.005), class=b, coef="dissim"),
          prior(cauchy(0,.1), class=sd),
          prior(cauchy(0,.1), class=sigma)),
  data=dhd3,
  sample_prior=TRUE,
  iter=1000,
  warmup=500,
  cores=parallel::detectCores())

save(fit_ce7, file="fit_ce7.rds", compress="xz")




fit_ce8 <- brm(
  bf(condeff ~ 0 + dissim*ori + (0 + dissim*ori||subject)),
  prior=c(prior(normal(0,.1), class=b, coef="ori0"),
          prior(normal(0,.1), class=b, coef="ori180"),
          prior(normal(0,.0005), class=b, coef="dissim"),
          prior(cauchy(0,.1), class=sd),
          prior(cauchy(0,.1), class=sigma)),
  data=dhd3,
  sample_prior=TRUE,
  iter=1000,
  warmup=500,
  cores=parallel::detectCores())

save(fit_ce8, file="fit_ce8.rds", compress="xz")


```

```{r load model etc. and tests, eval=FALSE}
load("fit_ce2.rds")

ce2BF <- 
  summary(fit_ce2, prob = 0.95)$fixed %>%
  data.frame %>%
  rownames_to_column(var = "Cond") %>%
  dplyr::select(-Bulk_ESS, -Tail_ESS, -Rhat) %>%
  magrittr::set_colnames(
    c("Cond", "Estimate", "SE", "Lower", "Upper")
  ) %>%
  mutate_if(is.numeric, ~ round(., 4))

ce2BF[4,2:5] <- ce2BF[1,2:5]+ce2BF[4,2:5]


ce2BF <- ce2BF %>% 
  # compute BF for each effect
  mutate(BF01 = hypothesis(fit_ce2, glue("{Cond} = 0") )$
      hypothesis$Evid.Ratio,
      BF10=1/BF01)

load("fit_ce3.rds")
ce3BF <- 
  summary(fit_ce3, prob = 0.95)$fixed %>%
  data.frame %>%
  rownames_to_column(var = "Cond") %>%
  dplyr::select(-Bulk_ESS, -Tail_ESS, -Rhat) %>%
  magrittr::set_colnames(
    c("Cond", "Estimate", "SE", "Lower", "Upper")
  ) %>%
  mutate_if(is.numeric, ~ round(., 4))

ce3BF[4,2:5] <- ce3BF[1,2:5]+ce3BF[4,2:5]


ce3BF <- ce3BF %>% 
  # compute BF for each effect
  mutate(BF01 = hypothesis(fit_ce3, glue("{Cond} = 0") )$
      hypothesis$Evid.Ratio,
      BF10=1/BF01)

load("fit_ce4.rds")

ce4BF <- 
  summary(fit_ce4, prob = 0.95)$fixed %>%
  data.frame %>%
  rownames_to_column(var = "Cond") %>%
  dplyr::select(-Bulk_ESS, -Tail_ESS, -Rhat) %>%
  magrittr::set_colnames(
    c("Cond", "Estimate", "SE", "Lower", "Upper")
  ) %>%
  mutate_if(is.numeric, ~ round(., 4))

ce4BF[4,2:5] <- ce4BF[1,2:5]+ce4BF[4,2:5]


ce4BF <- ce4BF %>% 
  # compute BF for each effect
  mutate(BF01 = hypothesis(fit_ce4, glue("{Cond} = 0") )$
      hypothesis$Evid.Ratio,
      BF10=1/BF01)

load("fit_ce5.rds")

ce5BF <- 
  summary(fit_ce5, prob = 0.95)$fixed %>%
  data.frame %>%
  rownames_to_column(var = "Cond") %>%
  dplyr::select(-Bulk_ESS, -Tail_ESS, -Rhat) %>%
  magrittr::set_colnames(
    c("Cond", "Estimate", "SE", "Lower", "Upper")
  ) %>%
  mutate_if(is.numeric, ~ round(., 4))

ce5BF[4,2:5] <- ce5BF[1,2:5]+ce5BF[4,2:5]


ce5BF <- ce5BF %>% 
  # compute BF for each effect
  mutate(BF01 = hypothesis(fit_ce5, glue("{Cond} = 0") )$
      hypothesis$Evid.Ratio,
      BF10=1/BF01)


load("fit_ce6.rds")

ce6BF <- 
  summary(fit_ce6, prob = 0.95)$fixed %>%
  data.frame %>%
  rownames_to_column(var = "Cond") %>%
  dplyr::select(-Bulk_ESS, -Tail_ESS, -Rhat) %>%
  magrittr::set_colnames(
    c("Cond", "Estimate", "SE", "Lower", "Upper")
  ) %>%
  mutate_if(is.numeric, ~ round(., 4))

ce6BF[4,2:5] <- ce6BF[1,2:5]+ce6BF[4,2:5]


ce6BF <- ce6BF %>% 
  # compute BF for each effect
  mutate(BF01 = hypothesis(fit_ce6, glue("{Cond} = 0") )$
      hypothesis$Evid.Ratio,
      BF10=1/BF01)

load("fit_ce7.rds")

ce7BF <- 
  summary(fit_ce7, prob = 0.95)$fixed %>%
  data.frame %>%
  rownames_to_column(var = "Cond") %>%
  dplyr::select(-Bulk_ESS, -Tail_ESS, -Rhat) %>%
  magrittr::set_colnames(
    c("Cond", "Estimate", "SE", "Lower", "Upper")
  ) %>%
  mutate_if(is.numeric, ~ round(., 4))

ce7BF[4,2:5] <- ce7BF[1,2:5]+ce7BF[4,2:5]


ce7BF <- ce7BF %>% 
  # compute BF for each effect
  mutate(BF01 = hypothesis(fit_ce7, glue("{Cond} = 0") )$
      hypothesis$Evid.Ratio) %>%
  mutate(BF01 = case_when(BF01 > 0 ~ BF01,
                          BF01 < 0 ~ 0),
      BF10=1/BF01)

load("fit_ce8.rds")
ce8BF <- 
  summary(fit_ce8, prob = 0.95)$fixed %>%
  data.frame %>%
  rownames_to_column(var = "Cond") %>%
  dplyr::select(-Bulk_ESS, -Tail_ESS, -Rhat) %>%
  magrittr::set_colnames(
    c("Cond", "Estimate", "SE", "Lower", "Upper")
  ) %>%
  mutate_if(is.numeric, ~ round(., 4))

ce8BF[4,2:5] <- ce8BF[1,2:5]+ce8BF[4,2:5]


ce8BF <- ce8BF %>% 
  # compute BF for each effect
  mutate(BF01 = hypothesis(fit_ce8, glue("{Cond} = 0") )$
      hypothesis$Evid.Ratio) %>%
  mutate(BF01 = case_when(BF01 > 0 ~ BF01,
                       BF01<0 ~ 0),
      BF10=1/BF01)

```

```{r Comparing ceBFs, eval=FALSE}
#ce3BF # This is the one with (0,10) priors
#ce4BF # This is the one with (0,1) priors
#ce5BF # This is the one with (.2, .1) and (0,.1) priors

ce3BF <- ce3BF %>% mutate(BF01=case_when(BF01 <0 ~ 0,
                                         BF01 >0 ~ BF01),
                          BF10=1/BF01)

ce4BF <- ce4BF %>% mutate(BF01=case_when(BF01 <0 ~ 0,
                                         BF01 >0 ~ BF01),
                          BF10=1/BF01)

ce5BF <- ce5BF %>% mutate(BF01=case_when(BF01 <0 ~ 0,
                                         BF01 >0 ~ BF01),
                          BF10=1/BF01)





```

```{r fig 6 final model, eval=FALSE}
fit_ce <- brm(
  bf(condeff ~ 0 + dissim*ori + (0 + dissim*ori||subject)),
  prior=c(prior(normal(0, 10), class = b),
          prior(cauchy(0, 10), class = sd),
          prior(cauchy(0, 10), class = sigma)),
  data=dhd3,
  sample_prior=TRUE,
  iter=2000,
  warmup=1000,
  cores = parallel::detectCores())


save(fit_ce, file="fit_ce.rds", compress="xz")

```

```{r fig 6 model data, cong eff}
load("fit_ce.rds")

ceBF <- 
  summary(fit_ce, prob = 0.95)$fixed %>%
  data.frame %>%
  rownames_to_column(var = "Cond") %>%
  dplyr::select(-Bulk_ESS, -Tail_ESS, -Rhat) %>%
  magrittr::set_colnames(
    c("Cond", "Estimate", "SE", "Lower", "Upper")
  ) %>%
  mutate_if(is.numeric, ~ round(., 4))

ceBF[4,2:5] <- ceBF[1,2:5]+ceBF[4,2:5]

ceBF <- ceBF %>% 
  # compute BF for each effect
  mutate(BF01 = hypothesis(fit_ce, glue("{Cond} = 0") )$
      hypothesis$Evid.Ratio) %>%
  mutate(BF01 = case_when(BF01 > 0 ~ BF01,
                          BF01 < 0 ~ 0),
      BF10=1/BF01)



```

```{r mcmc stuff - later, eval=FALSE}
mcmc_plot(fit_ce, type = 'rhat')

preddata <- crossing(
  dissim=seq(from=0, to = 100, length.out=101),
  ori = unique(dhd3$ori),
  subject = unique(dhd3$subject)
)

predicted_data <- add_fitted_draws(fit_ce, newdata = preddata, re_formula = NA, scale = "response")
predicted_data <- predicted_data %>% 
  mean_qi()

dhd3 %>%
  ggplot(aes(x = dissim, y = condeff, color = ori)) +
  stat_summary(fun.data = "mean_cl_normal") +
  geom_line(
    data = predicted_data,
    aes(x = dissim, y = .value, color = ori),
    inherit.aes = FALSE,
    size = 1.5,
    show.legend = FALSE
  ) +
  geom_ribbon(
    data = predicted_data,
    aes(
      x = dissim,
      ymin = .lower,
      ymax = .upper,
      fill = ori
    ),
    inherit.aes = FALSE,
    alpha = 0.3
  ) +
  #facet_grid(ROI~StimType) +
  #labs(x = "Exposure duration (ms)", y = "Averaged beta-values") +
  scale_color_manual(values = c("#057404", "#F8D482")) + 
  scale_fill_manual(values = c("#057404", "#F8D482"))
  #alxtheme

posterior_summary(fit_ce, pars="^b_")
postsum <- posterior_summary(fit_ce, pars="^b_")

```

```{r building a new model, eval=FALSE}


load("d_lohi.rda")

df <-  d_lohi12

df_trns <- df %>% mutate(iv2= case_when(task=="low"~log10(iv),
                                        TRUE~iv)) %>%
  group_by(task) %>%
  mutate(iv3=(iv2-mean(iv2))/sd(iv2),
         task=as.factor(task),
         guess=case_when(task=="low"~1,
                         task!="low"~0))

fit_condeff <- brm(
  bf(dv ~ 0.5*guess + (1-0.5*guess)*inv_logit(eta),
     eta ~ iv3*task*cond + (iv3*task*cond||subject), nl=TRUE),
  data=df_trns,
  family=bernoulli("identity"),
  prior=c(prior(normal(0,5), nlpar="eta"),
          prior(lognormal(0,.1), nlpar="eta", class="sd", group="subject")),
  cores=12)

save(fit_invlogit_final, file="fit_invlogit_final.rda", compress="xz")



logitBF <- 
  summary(fit_invlogit_final, prob = 0.95)$fixed %>%
  data.frame %>%
  rownames_to_column(var = "Cond") %>%
  dplyr::select(-Bulk_ESS, -Tail_ESS, -Rhat) %>%
  magrittr::set_colnames(
    c("Cond", "Estimate", "SE", "Lower", "Upper")
  ) %>%
  mutate_if(is.numeric, ~ round(., 4))

logit_BFs <- logitBF %>% 
  # compute BF for each effect
  mutate(BF01 = hypothesis(fit_invlogit_final, glue("{Cond} = 0") )$
      hypothesis$Evid.Ratio) %>%
  mutate(BF01 = case_when(BF01 > 0 ~ BF01,
                          BF01 < 0 ~ 0),
      BF10=1/BF01)




```



# Mannion replication (for Supplementary)

```{r mannion}

load("brmslsETA_1234th.rda")

b1 <- brmseta %>% select(location, subject, task, cond) %>%
  pivot_wider(names_from="cond", values_from="location") %>%
  mutate(ds = diff-same) %>% pivot_longer(3:6, names_to="cond", values_to="val") %>%
  pivot_wider(names_from=c("task","cond"), values_from="val") %>%
  mutate(high_ds = up_ds-inv_ds)

b2 <- b1 %>% select(subject, low_diff, low_same) %>% 
  pivot_longer(cols=2:3, names_to="context", values_to="values") %>%
  mutate(context=case_when(context=="low_diff"~ "Different",
                           context=="low_same"~ "Same")) 

b2 %>%
  ggplot(aes(x=context, y=values, group=subject))+
  xlab("Relative Orientation")+
  ylab("Contrast Threshold (Log & z)")+
  geom_boxplot(aes(group=context), size=1.2, width=.2)+
  stat_boxplot(geom ='errorbar',width=.1, size=1.2, aes(group=context))+
  geom_point(size=3)+
  geom_line(size=1.3, alpha=.25)+
  theme_classic()+
  muctheme+
  theme(strip.text=element_text(size=20),
        axis.text.x=element_text(size=14),
        axis.text.y=element_text(size=14),
        axis.title.y=element_text(size=20),
        axis.title.x=element_text(size=20))


load("fit_invlogit_1234th.rda")

hypothesis(fit_invlogit_1234th, "eta_iv3:tasklow = eta_iv3:tasklow:condsame")

```



# Correlations across tasks, residuals

```{r fig 8 correlations - preparations RofR method}

load("brmslsETA_1234th.rda")

b1 <- brmseta_1234 %>% select(location, subject, task, cond) %>%
  pivot_wider(names_from="cond", values_from="location") %>%
  mutate(ds = diff-same) %>% pivot_longer(3:6, names_to="cond", values_to="val") %>%
  pivot_wider(names_from=c("task","cond"), values_from="val") %>%
  mutate(high_ds = up_ds-inv_ds)

b2 <- b1

# Upright - Iso/Diff
uid <- lm(up_diff~up_iso, data=b2)
b2$uidpredicted <- predict(uid)
b2$uidresid <- residuals(uid)


# Upright - Iso/Same
uis <- lm(up_same~up_iso, data=b2)
b2$uispredicted <- predict(uis)
b2$uisresid <- residuals(uis)

# Inv - Iso/Diff
iid <- lm(inv_diff~inv_iso, data=b2)
b2$iidpredicted <- predict(iid)
b2$iidresid <- residuals(iid)

# Inv - Iso/Same
iis <- lm(inv_same~inv_iso, data=b2)
b2$iispredicted <- predict(iis)
b2$iisresid <- residuals(iis)

# Low - Iso/Diff
lid <- lm(low_diff~low_iso, data=b2)
b2$lidpredicted <- predict(lid)
b2$lidresid <- residuals(lid)

# Low - Iso/Same
lis <- lm(low_same~low_iso, data=b2)
b2$lispredicted <- predict(lis)
b2$lisresid <- residuals(lis)

b3 <- b2 %>% select(subject, uidresid, uisresid, iisresid, iidresid, lisresid, lidresid)

# Up - Diff ~ Same
uds <- lm(uidresid~uisresid, data=b3)
b3$udspredicted <- predict(uds)
b3$udsresid <- residuals(uds)

# Inv - Diff ~ Same
ids <- lm(iidresid~iisresid, data=b3)
b3$idspredicted <- predict(ids)
b3$idsresid <- residuals(ids)

# Low - Diff ~ Same
lds <- lm(lidresid~lisresid, data=b3)
b3$ldspredicted <- predict(lds)
b3$ldsresid <- residuals(lds)


# Up - Same ~ Diff 
usd <- lm(uisresid~uidresid, data=b3)
b3$usdpredicted <- predict(usd)
b3$usdresid <- residuals(usd)

# Inv - Same ~ Diff
isd <- lm(iisresid~iidresid, data=b3)
b3$isdpredicted <- predict(isd)
b3$isdresid <- residuals(isd)

# Low - Same ~ Diff 
lsd <- lm(lisresid~lidresid, data=b3)
b3$lsdpredicted <- predict(lsd)
b3$lsdresid <- residuals(lsd)


b4 <- b3 %>% select(subject, udsresid, idsresid, ldsresid, usdresid, isdresid, lsdresid, uisresid, uidresid, iisresid, iidresid, lisresid, lidresid)




```



```{r figures RofR}



uplowdisosisocor <- describe_posterior(correlationBF(b4$udsresid, b4$ldsresid))

uplowdisosisocor2 <- correlationBF(b4$udsresid, b4$ldsresid)

rofr1 <- b4 %>% ggplot(aes(x=udsresid, y=ldsresid)) +
  geom_point(size=3)+
  geom_smooth(method="lm", size=2, fullrange=TRUE)+
  ylab(expression(atop(bold("Contrast Detection"),atop("Contextual Modulation"))))+
  xlab(expression(atop(bold("Upright Face"),atop("Contextual Modulation"))))+
  ggthemes::scale_fill_colorblind() + 
  ggthemes::scale_color_colorblind() +
  theme_classic()+
  scale_x_continuous(limits=c(-3.27,3.27))+
  theme(axis.text.x = element_text(size=20),
        axis.text.y = element_text(size=20),
        axis.title=element_text(size=30),
        axis.title.y = element_text(vjust=-4),
        title=element_text(size=20),
        plot.title=element_text(size=30,hjust=0.5),
        plot.subtitle=element_text(size=20,hjust=0.5))+
  labs(title = expression(italic(r) == 0.15),
       subtitle = expression(italic(BF[10]) == 0.61))


invlowdisosisocor <- describe_posterior(correlationBF(b4$idsresid, b4$ldsresid))
invlowdisosisocor2 <- correlationBF(b4$idsresid, b4$ldsresid)

rofr2 <- b4 %>% ggplot(aes(x=idsresid, y=ldsresid)) +
  geom_point(size=3)+
  geom_smooth(method="lm", size=2, fullrange=TRUE)+
  ylab(expression(atop(bold("Contrast Detection"),atop("Contextual Modulation"))))+
  xlab(expression(atop(bold("Inverted Face"),atop("Contextual Modulation"))))+
  ggthemes::scale_fill_colorblind() + 
  scale_x_continuous(limits=c(-1,1))+
  ggthemes::scale_color_colorblind() +
  theme_classic()+
  theme(axis.text.x = element_text(size=20),
        axis.text.y = element_text(size=20),
        axis.title=element_text(size=30),
        axis.title.y = element_text(vjust=-4),
        title=element_text(size=20),
        plot.title=element_text(size=30,hjust=0.5),
        plot.subtitle=element_text(size=20,hjust=0.5))+
  labs(title = expression(italic(r) == 0.28),
       subtitle = expression(italic(BF[10]) == 4.58))


upinvdisosisocor <- describe_posterior(correlationBF(b4$udsresid, b4$idsresid))
upinvdisosisocor2 <- correlationBF(b4$udsresid, b4$idsresid)

rofr3 <- b4 %>% ggplot(aes(x=udsresid, y=idsresid)) +
  geom_point(size=3)+
  geom_smooth(method="lm", size=2, fullrange=TRUE)+
  ylab(expression(atop(bold("Inverted Face"),atop("Contextual Modulation"))))+
  xlab(expression(atop(bold("Upright Face"),atop("Contextual Modulation"))))+
  ggthemes::scale_fill_colorblind() + 
  ggthemes::scale_color_colorblind() +
  scale_x_continuous(limits=c(-3.27,3.27))+
  theme_classic()+
  theme(axis.text.x = element_text(size=20),
        axis.text.y = element_text(size=20),
        axis.title=element_text(size=30),
        axis.title.y = element_text(vjust=-4),
        title=element_text(size=20),
        plot.title=element_text(size=30,hjust=0.5),
        plot.subtitle=element_text(size=20,hjust=0.5))+
  labs(title = expression(italic(r) == 0.61),
       subtitle = expression(italic(BF[10]) == 455834))


rofr1 + rofr2 + rofr3

```

```{r fig 8 correlations - preparations Diff - Same method, eval=FALSE}


load("brmslsETA_1234th.rda")

b_ds <- brmseta_123 %>% select(location, subject, task, cond) %>%
  pivot_wider(names_from="cond", values_from="location")  %>%
  mutate(ds = diff-same) %>% pivot_longer(3:6, names_to="cond", values_to="val") %>%
  pivot_wider(names_from=c("task","cond"), values_from="val") %>%
  mutate(high_ds = up_ds-inv_ds) %>%
  select(subject, up_ds, inv_ds, low_ds)


uplowdscor <- describe_posterior(correlationBF(b_ds$up_ds, b_ds$low_ds))

uplowdscor2 <- correlationBF(b_ds$up_ds, b_ds$low_ds)

b_ds %>% ggplot(aes(x=up_ds, y=low_ds)) +
  geom_point(size=3)+
  geom_smooth(method="lm", size=2)+
  ylab("Grating")+
  xlab("Upright Face")+
  ggthemes::scale_fill_colorblind() + 
  ggthemes::scale_color_colorblind() +
  theme_classic()+
  annotate("text", -Inf, Inf, label = paste("italic(r) == ", as.character(round(uplowdscor$Median, 2)), sep=""), hjust = -.5, vjust = 3, size=15, parse=T)+
  annotate("text", -Inf, Inf, label = "italic(BF[10]) == 0.47", parse=T, hjust = -.2, vjust = 3.7, size=15)+
  #ggtitle("Upright vs. Low - (Diff~Iso)~(Same~Iso) residuals")+
  theme(axis.text.x = element_text(size=30),
        axis.text.y = element_text(size=30),
        axis.title=element_text(size=35),
        title=element_text(size=20))


invlowdscor <- describe_posterior(correlationBF(b_ds$inv_ds, b_ds$low_ds))
invlowdscor2 <- correlationBF(b_ds$inv_ds, b_ds$low_ds)

b_ds %>% ggplot(aes(x=inv_ds, y=low_ds)) +
  geom_point(size=3)+
  geom_smooth(method="lm", size=2)+
  ylab("Grating")+
  xlab("Inverted Face")+
  ggthemes::scale_fill_colorblind() + 
  ggthemes::scale_color_colorblind() +
  theme_classic()+
  annotate("text", -Inf, Inf, label = paste("italic(r) == ", as.character(round(invlowdscor$Median, 2)), sep=""), hjust = -.3, vjust = 3, size=15,parse=T)+
  annotate("text", -Inf, Inf, label = "italic(BF[10]) == 0.46",parse=T, hjust = -.1, vjust = 3.7, size=15)+
  #ggtitle("Inverted vs. Low - (Diff~Iso)~(Same~Iso) residuals")+
  theme(axis.text.x = element_text(size=30),
        axis.text.y = element_text(size=30),
        axis.title=element_text(size=35),
        title=element_text(size=20))

upinvdscor <- describe_posterior(correlationBF(b_ds$up_ds, b_ds$inv_ds))
upinvdscor2 <- correlationBF(b_ds$up_ds, b_ds$inv_ds)

b_ds %>% ggplot(aes(x=up_ds, y=inv_ds)) +
  geom_point(size=3)+
  geom_smooth(method="lm", size=2)+
  ylab("Inverted Face")+
  xlab("Upright Face")+
  ggthemes::scale_fill_colorblind() + 
  ggthemes::scale_color_colorblind() +
  theme_classic()+
  annotate("text", -Inf, Inf, label = paste("italic(r) == ", as.character(round(upinvdscor$Median, 2)), sep=""), hjust = -.2, vjust = 2.6, size=15, parse=T)+
  annotate("text", -Inf, Inf, label = "italic(BF[10]) == 308999",parse=T, hjust = -.05, vjust = 3.5, size=15)+
  #ggtitle("Upright vs. Inverted - (Diff~Iso)~(Same~Iso) residuals")+
  theme(axis.text.x = element_text(size=30),
        axis.text.y = element_text(size=30),
        axis.title=element_text(size=35),
        title=element_text(size=20))



```


# V1 Size Correlations
```{r V1 size correlations}

thresh_for_v1 <- b1 %>% filter(subject=="LH25" | subject=="LH15" | subject == "KEGI1911" | subject == "PEIT3004" | 
                            subject=="ERPA1905" | subject=="MONA1012" | subject == "AMNA3007" | subject == "LH26" |
                            subject == "MUMA1101" | subject=="MAVI1806" | subject=="SAAY1407" | subject == "CHPA3112" |
                            subject == "SIGA2401")

resids_for_v1 <- b2 %>% select(subject, uidresid, uisresid, iidresid, iisresid, lidresid, lisresid) %>%
  filter(subject=="LH25" | subject=="LH15" | subject == "KEGI1911" | subject == "PEIT3004" | 
                            subject=="ERPA1905" | subject=="MONA1012" | subject == "AMNA3007" | subject == "LH26" |
                            subject == "MUMA1101" | subject=="MAVI1806" | subject=="SAAY1407" | subject == "CHPA3112" |
                            subject == "SIGA2401")



v1_stuff <- b4 %>% filter(subject=="LH25" | subject=="LH15" | subject == "KEGI1911" | subject == "PEIT3004" | 
                            subject=="ERPA1905" | subject=="MONA1012" | subject == "AMNA3007" | subject == "LH26" |
                            subject == "MUMA1101" | subject=="MAVI1806" | subject=="SAAY1407" | subject == "CHPA3112" |
                            subject == "SIGA2401")
v1_sizes <- read.csv("out_file.txt", header=FALSE)
everything <- read.csv("cortical_stuff.csv")

everything <- as.tibble(everything) %>% select(-1)
everything <- everything %>% mutate(prop_left=lh_v1/lh_whole,
                                    prop_right=rh_v1/rh_whole)

wholes_v1s <- everything %>%
  mutate(sub=case_when(sub=="sub-01" ~ "LH25",
                       sub=="sub-02" ~ "LH15",
                       sub=="sub-03" ~ "KEGI1911",
                       sub=="sub-04" ~ "PEIT3004",
                       sub=="sub-09" ~ "ERPA1905",
                       sub=="sub-11" ~ "MONA1012",
                       sub=="sub-12" ~ "AMNA3007",
                       sub=="sub-17" ~ "LH26",
                       sub=="sub-18" ~ "MUMA1101",
                       sub=="sub-19" ~ "SIGA2401",
                       sub=="sub-21" ~ "CHPA3112",
                       sub=="sub-20" ~ "MAVI1806",
                       sub=="sub-22" ~ "SAAY1407")) %>%
  arrange(sub)

wholes_v1s <- wholes_v1s[1:13,] # This will change when we get Pauline and Eunice's behavioral data

wholes_v1s_behs <- cbind(wholes_v1s, v1_stuff$udsresid, v1_stuff$idsresid, v1_stuff$ldsresid) %>%
  rename(udsresid=`v1_stuff$udsresid`, idsresid=`v1_stuff$idsresid`, ldsresid=`v1_stuff$ldsresid`) %>% 
  mutate(total_prop = (prop_left+prop_right)/2,
         total_area = lh_v1 + rh_v1)



v1_sizes_long <- v1_sizes %>% rename(subject=V1, hemisphere=V2, size=V3) %>%
  pivot_wider(names_from=hemisphere, values_from=size) %>%
  mutate(subject=case_when(subject=="sub-01" ~ "LH25",
                           subject=="sub-02" ~ "LH15",
                           subject=="sub-03" ~ "KEGI1911",
                           subject=="sub-04" ~ "PEIT3004",
                           subject=="sub-09" ~ "ERPA1905",
                           subject=="sub-11" ~ "MONA1012",
                           subject=="sub-12" ~ "AMNA3007",
                           subject=="sub-17" ~ "LH26",
                           subject=="sub-18" ~ "MUMA1101",
                           subject=="sub-19" ~ "SIGA2401",
                           subject=="sub-20" ~ "MAVI1806",
                           subject=="sub-21" ~ "CHPA3112",
                           subject=="sub-22" ~ "SAAY1407"))
v1_sizes_long <- v1_sizes_long[order(v1_sizes_long$subject),]
v1_sizes_long <- v1_sizes_long[1:13,]

v1_beh <- cbind(v1_stuff, v1_sizes_long$lh, v1_sizes_long$rh)
v1_beh <- as.tibble(v1_beh)

v1_beh2 <- v1_beh %>% rename(lh=`v1_sizes_long$lh`, rh=`v1_sizes_long$rh`) %>%
  mutate(combined=lh+rh,
         new=case_when(subject == "CHPA3112" ~ "2",
                       subject == "SIGA2401"~ "1",
                       TRUE ~ "0")) 
  
upv1cor <- describe_posterior(correlationBF(v1_beh2$udsresid, v1_beh2$combined))
upv1cor2 <- correlationBF(v1_beh2$udsresid, v1_beh2$combined)
cor.test(v1_beh2$udsresid, v1_beh2$combined)
v1_beh2 %>% ggplot(aes(x=udsresid, y=combined)) + 
  geom_point(size=5, aes(color=new))+
  geom_smooth(method=lm, fullrange=TRUE)+
  theme_classic()+
  xlab("Upright Face Contextual Modulation Threshold Residual")+
  ylab("Left + Right Hemisphere Surface Area")+
  #xlim(500,1400)+
  #ylim(500,1400)+
  #ggtitle(bquote("Upright Faces Contextual vs. Functionally defined V1 surface area"~(mm^2)))+
  theme(plot.title= element_text(size=30, hjust=.5),
        axis.title.x = element_text(size=25),
        axis.title.y = element_text(size=25),
        axis.text.x = element_text(size=20),
        axis.text.y = element_text(size=20),
        plot.subtitle=element_text(size=20,hjust=0.5))+
  labs(title = expression(italic(r) == 0.41),
       subtitle = expression(italic(p) == .17))
        
invv1cor <- describe_posterior(correlationBF(v1_beh2$idsresid, v1_beh2$combined))
invv1cor2 <- correlationBF(v1_beh2$idsresid, v1_beh2$combined)
cor.test(v1_beh2$idsresid, v1_beh2$combined)
v1_beh2 %>% ggplot(aes(x=idsresid, y=combined)) + 
  geom_point(size=5, aes(color=new))+
  geom_smooth(method=lm, fullrange=TRUE)+
  theme_classic()+
  xlab("Inverted Face Contextual Modulation Threshold Residual")+
  ylab("Left + Right Hemisphere Surface Area")+
  #xlim(500,1400)+
  #ylim(500,1400)+
  #ggtitle(bquote("Upright Faces Contextual vs. Functionally defined V1 surface area"~(mm^2)))+
  theme(plot.title= element_text(size=30, hjust=.5),
        axis.title.x = element_text(size=25),
        axis.title.y = element_text(size=25),
        axis.text.x = element_text(size=20),
        axis.text.y = element_text(size=20),
        plot.subtitle=element_text(size=20,hjust=0.5))+
  labs(title = expression(italic(r) == 0.55),
       #subtitle = expression(italic(BF[10]) == 1.12))
       subtitle = expression(italic(p) == .05)) 

lowv1cor <- describe_posterior(correlationBF(v1_beh2$ldsresid, v1_beh2$combined))
lowv1cor2 <- correlationBF(v1_beh2$ldsresid, v1_beh2$combined)
cor.test(v1_beh3$ldsresid, v1_beh3$combined)
v1_beh2 %>% ggplot(aes(x=ldsresid, y=combined)) + 
  geom_point(size=5, aes(color=new))+
  geom_smooth(method=lm, fullrange=TRUE)+
  theme_classic()+
  xlab("Grating Contextual Modulation Threshold Residual")+
  ylab("Left + Right Hemisphere Surface Area")+
  #xlim(500,1400)+
  #ylim(500,1400)+
  #ggtitle(bquote("Upright Faces Contextual vs. Functionally defined V1 surface area"~(mm^2)))+
  theme(plot.title= element_text(size=30, hjust=.5),
        axis.title.x = element_text(size=25),
        axis.title.y = element_text(size=25),
        axis.text.x = element_text(size=20),
        axis.text.y = element_text(size=20),
        plot.subtitle=element_text(size=20,hjust=0.5))+
  labs(title = expression(italic(r) == 0.6),
       subtitle = expression(italic(p) == .03))


#write.csv(wholes_v1s_behs, file="v1s_behs.csv")


```

```{r plotting left and right superimposed}

v1_lr <- v1_beh2 %>% select(subject, udsresid, idsresid, ldsresid, lh, rh) %>% pivot_longer(5:6, names_to="hemi", values_to="area") %>%
  mutate(alpha=.15,
         udsresid = -udsresid,
         idsresid = -idsresid,
         ldsresid = -ldsresid)


v1_l <- v1_lr %>% filter(hemi=="lh")
v1_r <- v1_lr %>% filter(hemi=="rh")


v1_l %>% ggplot(aes(x=udsresid, y=area))+
  geom_smooth(color="darkblue", fill="darkblue", size=3, method=lm, fullrange=TRUE, alpha=.20)+
  geom_point(size=5, aes(color="darkblue"))+
  geom_smooth(aes(x=v1_r$udsresid, y=v1_r$area), color="orangered", fill="orangered", method=lm, fullrange=TRUE, alpha=.20, size=3)+
  geom_point(aes(x=v1_r$udsresid, y=v1_r$area, color="orangered"), size=5)+
  muctheme+
  xlab(bquote("Upright Face Contextual Modulation"))+
  ylab(bquote("V1 Surface Area"~(mm^2)))+
  ylim(400,1500)+
  xlim(-2.6,2.6)+
  scale_color_identity(name='Hemisphere',
                     breaks=c('darkblue', 'orangered'),
                     labels=c('Left V1' , 'Right V1'),
                     guide="legend")+
  annotate("text", -Inf, Inf, label = "italic(r) == -.38", hjust = -2.3, vjust = 1.5, size=10, parse=T, color="darkblue")+
  annotate("text", -Inf, Inf, label = "italic(BF[10]) == .72", parse=T, hjust = -2.25, vjust = 3.5, size=8, color="darkblue")+
  annotate("text", -Inf, Inf, label = "italic(r) == -.38", hjust = -3.5, vjust = 1.5, size=10, parse=T, color="orangered")+
  annotate("text", -Inf, Inf, label = "italic(BF[10]) == .73", parse=T, hjust = -3.4, vjust = 3.5, size=8, color="orangered")

v1_l %>% ggplot(aes(x=idsresid, y=area))+
  geom_smooth(color="darkblue", fill="darkblue", size=3, method=lm, fullrange=TRUE, alpha=.20)+
  geom_point(size=5, aes(color="darkblue"))+
  geom_smooth(aes(x=v1_r$idsresid, y=v1_r$area), color="orangered", fill="orangered", method=lm, fullrange=TRUE, alpha=.20, size=3)+
  geom_point(aes(x=v1_r$idsresid, y=v1_r$area, color="orangered"), size=5)+
  muctheme+
  xlab(bquote("Inverted Face Contextual Modulation"))+
  ylab(bquote("V1 Surface Area"~(mm^2)))+
  ylim(400,1500)+
  xlim(-.6,.6)+
  scale_color_identity(name='Hemisphere',
                     breaks=c('darkblue', 'orangered'),
                     labels=c('Left V1' , 'Right V1'),
                     guide="legend")+
  annotate("text", -Inf, Inf, label = "italic(r) == -.54", hjust = -2.3, vjust = 1.5, size=10, parse=T, color="darkblue")+
  annotate("text", -Inf, Inf, label = "italic(BF[10]) == 1.71", parse=T, hjust = -2.25, vjust = 3.5, size=8, color="darkblue")+
  annotate("text", -Inf, Inf, label = "italic(r) == -.48", hjust = -3.5, vjust = 1.5, size=10, parse=T, color="orangered")+
  annotate("text", -Inf, Inf, label = "italic(BF[10]) == 1.2", parse=T, hjust = -3.9, vjust = 3.5, size=8, color="orangered")

v1_l %>% ggplot(aes(x=ldsresid, y=area))+
  geom_smooth(color="darkblue", fill="darkblue", size=3, method=lm, fullrange=TRUE, alpha=.20)+
  geom_point(size=5, aes(color="darkblue"))+
  geom_smooth(aes(x=v1_r$ldsresid, y=v1_r$area), color="orangered", fill="orangered", method=lm, fullrange=TRUE, alpha=.20, size=3)+
  geom_point(aes(x=v1_r$ldsresid, y=v1_r$area, color="orangered"), size=5)+
  muctheme+
  xlab(bquote("Contrast Detection Contextual Modulation"))+
  ylab(bquote("V1 Surface Area"~(mm^2)))+
  ylim(400,1500)+
  xlim(-.35,.35)+
  scale_color_identity(name='Hemisphere',
                     breaks=c('darkblue', 'orangered'),
                     labels=c('Left V1' , 'Right V1'),
                     guide="legend")+
  annotate("text", -Inf, Inf, label = "italic(r) == -.56", hjust = -2.3, vjust = 1.5, size=10, parse=T, color="darkblue")+
  annotate("text", -Inf, Inf, label = "italic(BF[10]) == 2.04", parse=T, hjust = -2.25, vjust = 3.5, size=8, color="darkblue")+
  annotate("text", -Inf, Inf, label = "italic(r) == -.56", hjust = -3.5, vjust = 1.5, size=10, parse=T, color="orangered")+
  annotate("text", -Inf, Inf, label = "italic(BF[10]) == 2.03", parse=T, hjust = -3.45, vjust = 3.5, size=8, color="orangered")

thresh_for_v1 <- thresh_for_v1 %>% mutate(totv1 = lh_v1+rh_v1)

#write.csv(thresh_for_v1, file="resids_threshs_v1.csv")
#save(thresh_for_v1, file="regs_threshs_v1.rda")

```


```{r checking whether those numbers deviate from the norm}
b5 <- b4 %>%
  mutate(Scan=case_when(subject=="LH25" ~ "Scanned",
                           subject=="LH15" ~ "Scanned",
                           subject=="KEGI1911" ~ "Scanned",
                           subject=="PEIT3004" ~ "Scanned",
                           subject=="ERPA1905" ~ "Scanned",
                           subject=="MONA1012" ~ "Scanned",
                           subject=="AMNA3007" ~ "Scanned",
                           subject=="LH26" ~ "Scanned",
                           subject=="MUMA1101" ~ "Scanned",
                           subject=="CHPA3112" ~ "Scanned",
                           subject=="SIGA2401" ~ "Scanned",
                           TRUE ~ "Not-Scanned"))


b5 %>% ggplot(aes(y="ldsresid",x=ldsresid)) +
  geom_boxplot(fill="grey",outlier.shape=NA, alpha=.5)+
  geom_point(position="jitter",size=4,aes(color=Scan, alpha=Scan, size=Scan))+
  scale_color_manual(values=c("black", "red"))+
  scale_alpha_manual(values=c(1,1))+
  scale_size_manual(values=c(2,25))+
  theme_classic()+
  theme(axis.title.y = element_blank(),
        axis.text.y = element_blank(),
        axis.title.x = element_text(size=25),
        axis.text.x = element_text(size=15))+
  xlab("Grating Contextual Modulation Threshold Residual")



```

# Vectoral Correlations

```{r Veccor}
load("brmslsETA_1234th.rda")
library(DescTools)

b1 <- brmseta_1234 %>% select(location, subject, task, cond) %>%
  pivot_wider(names_from="cond", values_from="location") %>%
  mutate(ds = diff-same) %>% pivot_longer(3:6, names_to="cond", values_to="val") %>%
  pivot_wider(names_from=c("task","cond"), values_from="val") %>%
  mutate(high_ds = up_ds-inv_ds)

vecs <- b1 %>% select(subject, inv_same, inv_iso, inv_diff, low_same, 
                      low_iso, low_diff, up_same, up_iso, up_diff)
vecs2 <- data.frame(vecs)



veccors <- c()
veccorsFisherZ <- c()
for(i in seq(1:nrow(vecs))){
  up <- vecs[i,8:10]
  low <- vecs[i,5:7]
  inv <- vecs[i,2:4]
  
  up <- up %>% pivot_longer(1:3,names_to="names",values_to="values")
  inv <- inv %>% pivot_longer(1:3, names_to="names", values_to="values")
  low <- low %>% pivot_longer(1:3, names_to="names", values_to="values")
  
  uplowcurr <- cor(up$values, low$values, method="pearson")
  invlowcurr <- cor(inv$values, low$values,method="pearson")
  upinvcurr <- cor(up$values, inv$values, method="pearson")
  
  uplowFZ <- FisherZ(uplowcurr)
  invlowFZ <- FisherZ(invlowcurr)
  upinvFZ <- FisherZ(upinvcurr)
  
  curr <- tibble(subject = vecs$subject[i], uplow = uplowcurr, invlow=invlowcurr, upinv=upinvcurr)
  
  currZ <- tibble(subject = vecs$subject[i], uplowFZ = uplowFZ, invlowFZ=invlowFZ, upinvFZ=upinvFZ)
  
  veccors <- rbind(veccors, curr)
  veccorsFisherZ <- rbind(veccorsFisherZ, currZ)
  
  
}


veccors2 <- veccors %>% pivot_longer(2:4, names_to="corrcoef", values_to="val") %>%
  mutate(corrcoef = case_when(corrcoef=="invlow" ~ "Inv vs. Grating",
                              corrcoef=="uplow" ~ "Up vs. Grating",
                              corrcoef=="upinv" ~ "Up vs. Inv"))


veccors2$corrcoef <- ordered(veccors2$corrcoef, levels = c("Up vs. Grating", "Inv vs. Grating","Up vs. Inv"))

veccors2 %>% ggplot(aes(x=corrcoef, y=val))+
  geom_boxplot(fill="goldenrod", width=.25, outlier.shape=NA)+
  stat_boxplot(geom="errorbar", width=.15)+
  geom_point(position="jitter", fill="black", color="midnightblue")+
  theme_classic()+
  xlab("")+
  ylab("Pearson Correlation Coefficient")+
  theme(legend.position="none")+
  geom_hline(yintercept=0)+
  muctheme+
  theme(axis.text.x = element_text(size=12),
        axis.text.y = element_text(size=20),
        axis.title=element_text(size=50),
        axis.title.y=element_text(size=25))+
  scale_x_discrete(labels = c("Upright Face\n vs\n Contrast Detection", 
                              "Inverted Face\n vs\n Contrast Detection",
                              "Upright Face\n vs\n Inverted Face"))


# Statistical values

UpBas <- veccors2 %>% filter(corrcoef == "Up vs. Grating")
InvBas <- veccors2 %>% filter(corrcoef == "Inv vs. Grating")
UpInv <- veccors2 %>% filter(corrcoef == "Up vs. Inv")

BFUpBas <- ttestBF(UpBas$val)
UpBasavgR <- mean(UpBas$val)
BFInvBas <- ttestBF(InvBas$val)
InvBasavgR <- mean(InvBas$val)
BFUpInv <- ttestBF(UpInv$val)
UpInvavgR <- mean(UpInv$val)

vecs3 <- vecs %>% pivot_longer(2:10, names_to="name",values_to="val")
vecs4 <- vecs3 %>% separate(name, into=c("task","cond"), sep="_") %>%
  mutate(jit1 = case_when(cond=="diff"~1,
                          cond=="iso"~3,
                          cond=="same"~2),
         jit2 = jitter(jit1),
         task=case_when(task=="inv"~"Inverted Face",
                        task=="up"~"Upright Face",
                        task=="low"~"Contrast Detection"),
         cond=case_when(cond=="diff"~"Different",
                        cond=="same"~"Same",
                        cond=="iso"~"Isolated"))


vecsmean <- vecs4 %>% group_by(task, cond) %>%
  summarise(val=mean(val))

vecsmean$task <- ordered(vecsmean$task, levels = c("Upright Face", "Inverted Face","Contrast Detection"))
vecsmean$cond <- ordered(vecsmean$cond, levels = c("Different", "Same","Isolated"))
vecsmean <- arrange(vecsmean, cond)

vecs4$task <- ordered(vecs4$task, levels = c("Upright Face", "Inverted Face","Contrast Detection"))
vecs4$cond <- ordered(vecs4$cond, levels = c("Different", "Same","Isolated"))



vecs4 %>% ggplot(aes(x=jit2, y=val, color=cond))+
  geom_boxplot(aes(x=cond,fill=cond), alpha=.5, outlier.shape=NA)+
  geom_point(aes(group=subject),size=5)+
  geom_line(aes(group=subject), color="gray", alpha=.3)+
  facet_wrap(~task)+
  stat_boxplot(aes(group=cond),geom="errorbar", width=.1)+
  theme_classic()+
  ggthemes::scale_fill_colorblind() + 
  ggthemes::scale_color_colorblind() +
  ylab("Threshold")+
  xlab("Condition")+
  geom_point(data=vecsmean, aes(x=cond, y=val), color="black", alpha=.6, size=4)+
  muctheme+
  theme(strip.text=element_text(size=40),
        axis.text.x=element_text(size=0),
        axis.text.y=element_text(size=28),
        axis.title.y=element_text(size=40),
        axis.title.x=element_text(size=0),
        legend.text=element_text(size=40))+
  geom_path(data=vecsmean, aes(x=cond, y=val, group=task), color="black", alpha=.6,size=2)



```


# LOO Comparison and anovaBF

```{r model comparison output}
load("LOOcomp_m1_m12_m13_m123.rda")

LOOCV
```

```{r anovaBF}
load("brmslsETA_final.rda")

brmseta2 <- brmseta %>% 
  mutate(task = as.factor(task),
         cond = as.factor(cond))

int_anova <- anovaBF(location ~ task*cond, data=brmseta2)

int_anova@bayesFactor


```

```{r anova BRMS, eval=FALSE}

load("brmslsETA_123rd.rda")

brmseta2 <- brmseta_123 %>% 
  mutate(task = as.factor(task),
         cond = as.factor(cond))

fit_thr_anova <- brm(
  bf(location ~ 0 + task*cond + (0 + task*cond||subject)),
  data=brmseta2,
  prior=c(prior(normal(0, 1), class = b),
  prior(cauchy(0, 10), class = sd),
  prior(cauchy(0, 10), class = sigma)),
  cores=12,
  warmup = 2000, # 2000 first iterations will be set aside.
  iter   = 4000,
  chains = 4)
save(fit_thr_anova, file="fit_thr_anova_123.rda", compress="xz")


```

## BRMS Model - General ANOVA

```{r brms general ANOVA, eval=FALSE}

load("fit_invlogit_final.rda")

md <- fit_invlogit_final

d <- fixef(md, summary=T)

describe_posterior(md)
plot(conditional_effects(md))

conds <- unique(d_lohi12$cond)
marginal_effects(md, "iv3:task", conditions=conds )


conditional_effects(md, conditions="iv3:task:cond")

conditions <- make_conditions(md, "iv3")
conditional_effects(md, "task:cond", conditions = conditions)
## only include points close to the specified values of 
me3int <- conditional_effects(
  md, "task:cond", conditions = conditions,
  select_points = 0.1
)
plot(me, points = TRUE)

posterior_samples(md, "^b_eta")

metask <- conditional_effects(md, "task")
mecond <- conditional_effects(md, "cond")
inttaskcond <- conditional_effects(md, "task:cond")

lmetask <- metask$task %>% select(estimate__, se__, lower__, upper__, effect1__) %>% 
  pivot_longer()

conditions <- data.frame(iv3 = c(0, 1))
plot(conditional_effects(md, effects = "task:cond", 
                         conditions = conditions))

bayesfactor_parameters(md, null=0)

df <- d_lohi12

df_trns <- df %>% mutate(iv2=case_when(task=="low"~log10(iv),
                                       TRUE~iv)) %>%
  group_by(task) %>%
  mutate(iv3=(iv2-mean(iv2))/sd(iv2),
         task=as.factor(task))

load("fit_invlogit_final.rda")
pred_df <- expand.grid(iv3 = seq(-3,3, l = 101),
                       task = unique(df_trns$task),
                       cond = unique(df_trns$cond))
pred_df <- pred_df %>% mutate(guess=case_when(task=="low"~1,
                                              task!="low"~0))
pred_df <- bind_cols(pred_df, as_tibble(fitted(fit_invlogit_final, newdata = pred_df, 
                                               re_formula = NA,scale="linear")))

pred_df %>% ggplot(aes(x=iv3, y=Estimate, color=cond))+
  geom_line()+
  geom_ribbon(aes(ymin=Q2.5, ymax=Q97.5, fill=cond), alpha=.1, color="NA")+
  ggthemes::scale_fill_colorblind() + 
  ggthemes::scale_color_colorblind() + 
  geom_hline(yintercept=.75)+
  facet_wrap(~task)+
  theme_classic()+
  #theme(legend.position="none")+
  #theme(strip.text.x = element_text(size = 8, colour = "orange", angle = 90))
  #theme(strip.text.x = element_blank(),
        #strip.background = element_blank())+
  theme(axis.text.x = element_text(size=15),
        axis.text.y = element_text(size=15),
        strip.text = element_text(size=25))



```

```{r model comparison add LOO, eval=FALSE}
#iv3 # fit_m3
#iv3*task # fit_int3
#iv3*cond # fit_int2
#iv3*task*cond #fit_tripint

load("Meff_int_models/fit_null.rda")
load("Meff_int_models/fit_null_sub.rda")
load("Meff_int_models/fit_m3.rda")
load("Meff_int_models/fit_int3.rda")
load("Meff_int_models/fit_int2.rda")
load("Meff_int_models/fit_tripint.rda")

# performing model comparison via approximate LOO-CV
fit_null <- add_criterion(fit_null, "loo")
fit_null_sub <- add_criterion(fit_null_sub, "loo")
fit_m3 <- add_criterion(fit_m3, "loo")
fit_int3 <- add_criterion(fit_int3, "loo")
fit_int2 <- add_criterion(fit_int2, "loo")
fit_tripint <- add_criterion(fit_tripint, "loo")


save(fit_null, file="Meff_int_models/fit_null.rda")
save(fit_null_sub, file="Meff_int_models/fit_null_sub.rda")
save(fit_m3, file="Meff_int_models/fit_m3.rda")
save(fit_int2, file="Meff_int_models/fit_int2.rda")
save(fit_int3, file="Meff_int_models/fit_int3.rda")
save(fit_tripint, file="Meff_int_models/fit_tripint.rda")

```

```{r LOOCV, eval=FALSE}

load("Meff_int_models/fit_null.rda")
load("Meff_int_models/fit_null_sub.rda")
load("Meff_int_models/fit_m1.rda")
load("Meff_int_models/fit_m2.rda")
load("Meff_int_models/fit_m3.rda")
load("Meff_int_models/fit_2mef1.rda")
load("Meff_int_models/fit_2mef2.rda")
load("Meff_int_models/fit_2mef3.rda")
load("Meff_int_models/fit_meff1_int.rda")
load("Meff_int_models/fit_meff2_int.rda")
load("Meff_int_models/fit_meff3_int.rda")
load("Meff_int_models/fit_3mef.rda")
load("Meff_int_models/fit_int1.rda")
load("Meff_int_models/fit_int2.rda")
load("Meff_int_models/fit_int3.rda")
load("Meff_int_models/fit_tripint.rda")





# displays the model comparison table
LOOCV <- loo_compare(fit_null, fit_null_sub, fit_m1, fit_m2, fit_m3, fit_2mef1, fit_2mef2,
                     fit_2mef3, fit_3mef, fit_int1, fit_int2, fit_int3, fit_meff1_int, fit_meff2_int, fit_meff3_int, fit_tripint,
  criterion = "loo"
  )

# LOOCV <- loo_compare(fit_null_sub, fit_m3,
#   criterion = "loo"
#   )

LOOCV2 <- as_tibble(LOOCV, rownames="model")

LOOCV2 <- LOOCV2 %>% mutate(formula=case_when(model=="fit_tripint"~"iv3*task*cond||sub",
                                              model=="fit_meff1_int"~"cond+iv3*task ||sub",
                                              model=="fit_meff2_int"~"task+iv3*cond ||sub",
                                              model=="fit_meff3_int"~"iv3+task*cond ||sub",
                                              model=="fit_3mef"~"iv3+task+cond||sub",
                                              model=="fit_int1"~"cond*task||sub",
                                              model=="fit_int2"~"cond*iv3||sub",
                                              model=="fit_int3"~"iv3*task||sub",
                                              model=="fit_2mef1"~"cond+task ||sub",
                                              model=="fit_2mef2"~"cond+iv3 ||sub",
                                              model=="fit_2mef3"~"iv3+task ||sub",
                                              model=="fit_m1"~"cond||sub",
                                              model=="fit_m2"~"task||sub",
                                              model=="fit_m3"~"iv3||sub",
                                              model=="fit_null_sub"~"1||sub",
                                              model=="fit_null"~"1"))
LOOCV2$formula <- factor(LOOCV2$formula, levels=LOOCV2$formula)


LOOCV2 %>% ggplot(aes(x=formula, y=elpd_diff))+
  geom_point()+
  geom_errorbar(aes(ymax=elpd_diff+se_diff, ymin=elpd_diff-se_diff))
```



# Thresholds - CrIs

```{r anova BRMS model details, credible intervals}


load("fit_thr_anova_123.rda")

fe <- as_tibble(fixef(fit_thr_anova, summary=F)) %>% 
  mutate(iter = 1:nrow(.)) %>% 
  rename(`taskinv:conddiff`=taskinv,
         `tasklow:conddiff`=tasklow,
         `taskup:conddiff`=taskup,
         `taskinv:condiso`=condiso,
         `taskinv:condsame`=condsame) %>%
  pivot_longer(cols = 1:9, names_to = "param", values_to = "est") %>% 
  separate(param, into = c("task", "condition"), sep = ":") %>%
  mutate(condition = stringr::str_remove(condition, pattern = "cond"),
         task = stringr::str_remove(task, pattern = "task"))

cond_eff <- fe %>% pivot_wider(names_from = "condition", values_from = "est") %>%
  mutate(diff_iso = diff - iso, 
         diff_same = diff - same,
         same_iso = same - iso)

cond_eff2 <- cond_eff %>% select(iter, task, diff_iso, diff_same, same_iso) %>% 
  pivot_longer(cols = diff_iso:same_iso, names_to = "difference", values_to = "value") %>% 
  group_by(task,  difference) %>% 
  mutate(sig = (quantile(value, .025) < 0 & quantile(value, .975) < 0) | 
           (quantile(value, .025) > 0 & quantile(value, .975) > 0), 
         sig_alpha = ifelse(sig == T, .5, .3)) #%>% 

cond_eff2 %>% ggplot(aes(x = value, color = difference, fill = difference)) + 
  geom_density(aes(alpha = sig_alpha)) + 
  geom_vline(xintercept = 0, linetype = "dashed") + 
  facet_wrap(~ task) + 
  ggthemes::scale_color_colorblind() + ggthemes::scale_fill_colorblind() + 
  guides(alpha = F) + 
  ggtitle("Posterior differences between all contexts", 
          subtitle = "Different tasks")


post <- posterior_samples(fit_thr_anova, pars = "^b_*")

up_ds <- as.data.frame(cond_eff %>% filter(task=="up") %>% select(diff_same, same_iso, diff_iso))
inv_ds <- cond_eff %>% filter(task=="inv") %>% select(diff_same, same_iso, diff_iso)
low_ds <- cond_eff %>% filter(task=="low") %>% select(diff_same, same_iso, diff_iso)


beta_up_ds <- median(up_ds$diff_same)
credint_up_ds <- ci(up_ds, method="HDI")

beta_inv_ds <- median(inv_ds$diff_same)
credint_inv_ds <- ci(inv_ds, method="HDI")

beta_low_ds <- median(low_ds$diff_same)
credint_low_ds <- ci(low_ds, method="HDI")


beta_up_si <- median(up_ds$same_iso)
credint_up_si <- ci(up_ds$same_iso, method="HDI")

beta_inv_si <- median(inv_ds$same_iso)
credint_inv_si <- ci(inv_ds$same_iso, method="HDI")

beta_low_si <- median(low_ds$same_iso)
credint_low_si <- ci(low_ds$same_iso, method="HDI")


beta_up_di <- median(up_ds$diff_iso)
credint_up_di <- ci(up_ds$diff_iso, method="HDI")

beta_inv_di <- median(inv_ds$diff_iso)
credint_inv_di <- ci(inv_ds$diff_iso, method="HDI")

beta_low_di <- median(low_ds$diff_iso)
credint_low_di <- ci(low_ds$diff_iso, method="HDI")







```



# Supplementary Figures

## PsychFits
```{r PsychFitEyeMatching model, eval=FALSE}
load("d_lohi123.rda")

dfh <- d_lohi123 %>% filter(task != "low") %>%
    mutate(task = case_when(task=="upright"~"Upright Face",
                          task=="inverted"~"Inverted Face"),
         cond = case_when(cond=="same"~"Same",
                          cond=="diff"~"Different",
                          cond=="iso"~"Isolated")) %>% 
  rename(Condition=cond, Orientation=task)




mod <- glm(dv ~ iv * Orientation * Condition ,binomial(),dfh)
mdata <- expand.grid(iv= seq(0,100, l = 1001),
                     Orientation=c("Upright Face","Inverted Face"),
                     Condition=c("Same","Different","Isolated"))
mdata <- add_column(mdata, fit = predict(mod, newdata = mdata, type = 'response'))
fam <- family(mod)
ilink <- fam$linkinv
ilink <- family(mod)$linkinv
mdata <- bind_cols(mdata, setNames(as_tibble(predict(mod, mdata, se.fit = TRUE)[1:2]),
                                   c('fit_link','se_link')))
mdata <- mutate(mdata,
                fit_resp  = ilink(fit_link),
                right_upr = ilink(fit_link + (2 * se_link)),
                right_lwr = ilink(fit_link - (2 * se_link)))


dfh_avg <- dfh %>% group_by(iv, Condition, Orientation) %>%
  summarise(mean=mean(dv),
            sd=sd(dv),
            se=sd/sqrt(length(unique(dfh$subject))))# %>%


dfh_avg$Orientation <- ordered(dfh_avg$Orientation, levels = c("Upright Face", "Inverted Face"))
dfh_avg$Condition <- ordered(dfh_avg$Condition, levels = c("Different","Same","Isolated"))

dfh$Orientation <- ordered(dfh$Orientation, levels = c("Upright Face", "Inverted Face"))
dfh$Condition <- ordered(dfh$Condition, levels = c("Different","Same","Isolated"))

mdata %>% ggplot(aes(x=iv, y=fit, group=Condition, color=Condition, fill=Condition))+
    geom_line(size=1.3, show.legend=TRUE, alpha=1)+
    geom_point(size=5, data=dfh_avg, aes(x=iv, y=mean), alpha=.5)+
    geom_errorbar(data=dfh_avg, inherit.aes=F, aes(x=iv, ymax=mean+se, ymin=mean-se), width=2.5, alpha=.3)+
    geom_ribbon(aes(ymin=right_lwr, ymax=right_upr), alpha=.25)+
    facet_wrap(~Orientation)+
    theme_classic() + 
    ggthemes::scale_color_colorblind()+
    ggthemes::scale_fill_colorblind()+
    scale_x_continuous(breaks=c(0,24,36,53,80))+
    scale_y_continuous(limits=c(0,1))+
    xlab("Dissimilarity (%)")+
    ylab("Proportion of 'Different' Responses")+
    theme(strip.background=element_blank(),
          strip.text=element_text(size=25),
          axis.title=element_text(size=20),
          axis.text=element_text(size=15))



     




```



## Acc Figures

```{r Acc Figures for HP demonstration}
load("hi_clean1234.rda")

Nsub <- nrow(hi_clean1234 %>% group_by(subject) %>% tally())

hi_df2 <- hi_clean1234 %>% group_by(ori, dissim, cong) %>% 
  summarise(meanacc=mean(acc), sd=sd(acc), se=sd/sqrt(Nsub)) %>%
  mutate(dissim=as.factor(dissim),
         cong=as.factor(cong),
         ori=as.factor(ori),
         color=case_when(cong=="cong"~"#00b159",
                         cong=="incong"~"#d11141",
                         cong=="iso"~"#cccccc")) %>%
  mutate(ori=case_when(ori==0~"Upright",
                       ori==180~"Inverted")) %>%
  rename(Orientation=ori, Congruency=cong, Dissimilarity=dissim, Accuracy=meanacc) %>%
  mutate(Congruency=case_when(Congruency=="cong" ~ "Congruent",
                              Congruency=="incong" ~ "Incongruent",
                              Congruency=="iso" ~ "Isolated"))


hi_df2$Orientation <- ordered(hi_df2$Orientation, levels = c("Upright", "Inverted"))

write.csv(hi_df2, "hi_df2.csv")


colors=c("Congruent"="green4",
         "Incongruent"="orangered3",
         "Isolated"="grey60")

diss0 <- hi_df2 %>% filter(Dissimilarity==0) %>% ggplot(aes(y=Accuracy, x=Congruency, fill=Congruency))+
  geom_bar(stat='identity')+
  geom_errorbar(aes(ymax=Accuracy+se, ymin=Accuracy-se), position=position_dodge(0.9), width=.1)+
  scale_fill_manual(values=colors)+
  facet_wrap(~Orientation, strip.position="bottom")+
  labs(title="Target Dissimilarity = 0%")+
  theme(strip.text=element_text(size=25),
        strip.background=element_blank(),
        axis.title=element_text(size=20),
        axis.text = element_text(size=15),
        axis.title.x=element_blank(),
        axis.text.x = element_blank(),
        plot.title=element_text(size=25, hjust=0.5),
        legend.position = "none")

diss24 <- hi_df2 %>% filter(Dissimilarity==24) %>% ggplot(aes(y=Accuracy, x=Congruency, fill=Congruency))+
  geom_bar(stat='identity')+
  geom_errorbar(aes(ymax=Accuracy+se, ymin=Accuracy-se), position=position_dodge(0.9), width=.1)+
  scale_fill_manual(values=colors) +
  facet_wrap(~Orientation, strip.position="bottom")+
  labs(title="Target Dissimilarity = 24%")+
  theme(strip.text=element_text(size=25),
        strip.background=element_blank(),
        axis.title=element_text(size=20),
        axis.text = element_text(size=15),
        axis.title.x=element_blank(),
        axis.text.x = element_blank(),
        plot.title=element_text(size=25, hjust=0.5),
        legend.position = "none")

diss36 <- hi_df2 %>% filter(Dissimilarity==36) %>% ggplot(aes(y=Accuracy, x=Congruency, fill=Congruency))+
  geom_bar(stat='identity')+
  geom_errorbar(aes(ymax=Accuracy+se, ymin=Accuracy-se), position=position_dodge(0.9), width=.1)+
  scale_fill_manual(values=colors) +
  facet_wrap(~Orientation, strip.position="bottom")+
  labs(title="Target Dissimilarity = 36%")+
  theme(strip.text=element_text(size=25),
        strip.background=element_blank(),
        axis.title=element_text(size=20),
        axis.text = element_text(size=15),
        axis.title.x=element_blank(),
        axis.text.x = element_blank(),
        plot.title=element_text(size=25, hjust=0.5))

diss53 <- hi_df2 %>% filter(Dissimilarity==53) %>% ggplot(aes(y=Accuracy, x=Congruency, fill=Congruency))+
  geom_bar(stat='identity')+
  geom_errorbar(aes(ymax=Accuracy+se, ymin=Accuracy-se), position=position_dodge(0.9), width=.1)+
  scale_fill_manual(values=colors) +
  facet_wrap(~Orientation, strip.position="bottom")+
  labs(title="Target Dissimilarity = 53%")+
  theme(strip.text=element_text(size=25),
        strip.background=element_blank(),
        axis.title=element_text(size=20),
        axis.text = element_text(size=15),
        axis.title.x=element_blank(),
        axis.text.x = element_blank(),
        plot.title=element_text(size=25, hjust=0.5),
        legend.position = "none")

diss80 <- hi_df2 %>% filter(Dissimilarity==80) %>% ggplot(aes(y=Accuracy, x=Congruency, fill=Congruency))+
  geom_bar(stat='identity')+
  geom_errorbar(aes(ymax=Accuracy+se, ymin=Accuracy-se), position=position_dodge(0.9), width=.1)+
  scale_fill_manual(values=colors) +
  facet_wrap(~Orientation, strip.position="bottom")+
  labs(title="Target Dissimilarity = 80%")+
  theme(strip.text=element_text(size=25),
        strip.background=element_blank(),
        axis.title=element_text(size=20),
        axis.text = element_text(size=15),
        axis.title.x=element_blank(),
        axis.text.x = element_blank(),
        plot.title=element_text(size=25, hjust=0.5),
        legend.position = "none")

diss0 + diss24 + diss36 + diss53 + diss80



```



## Psych Function Raw Data Face Task
```{r raw data psych func}
load("hi_clean123.rda")

Nsub <- nrow(hi_clean123 %>% group_by(subject) %>% tally())

hi_df3 <- hi_clean123 %>% group_by(ori, dissim, cong) %>% 
  summarise(meandiff=mean(diff), sd=sd(diff), se=sd/sqrt(Nsub)) %>%
  mutate(dissim=as.factor(dissim),
         cong=as.factor(cong),
         ori=as.factor(ori),
         color=case_when(cong=="cong"~"#00b159",
                         cong=="incong"~"#d11141",
                         cong=="iso"~"#cccccc")) %>%
  mutate(ori=case_when(ori==0~"Upright",
                       ori==180~"Inverted")) %>%
  rename(Orientation=ori, Congruency=cong, Dissimilarity=dissim, PropDiff=meandiff) %>%
  mutate(Congruency=case_when(Congruency=="cong" ~ "Congruent",
                              Congruency=="incong" ~ "Incongruent",
                              Congruency=="iso" ~ "Isolated"))


hi_df2$Orientation <- ordered(hi_df2$Orientation, levels = c("Upright", "Inverted"))


dfh <- d_lohi123 %>% filter(task != "low") %>%
    mutate(task = case_when(task=="upright"~"Upright Face",
                          task=="inverted"~"Inverted Face"),
         cond = case_when(cond=="same"~"Same",
                          cond=="diff"~"Different",
                          cond=="iso"~"Isolated")) %>% 
  rename(Condition=cond, Orientation=task)




mod <- glm(dv ~ iv * Orientation * Condition ,binomial(),dfh)
mdata <- expand.grid(iv= seq(0,100, l = 1001),
                     Orientation=c("Upright Face","Inverted Face"),
                     Condition=c("Same","Different","Isolated"))
mdata <- add_column(mdata, fit = predict(mod, newdata = mdata, type = 'response'))
fam <- family(mod)
ilink <- fam$linkinv
ilink <- family(mod)$linkinv
mdata <- bind_cols(mdata, setNames(as_tibble(predict(mod, mdata, se.fit = TRUE)[1:2]),
                                   c('fit_link','se_link')))
mdata <- mutate(mdata,
                fit_resp  = ilink(fit_link),
                right_upr = ilink(fit_link + (2 * se_link)),
                right_lwr = ilink(fit_link - (2 * se_link)))


dfh_avg <- dfh %>% group_by(iv, Condition, Orientation) %>%
  summarise(mean=mean(dv),
            sd=sd(dv),
            se=sd/sqrt(length(unique(dfh$subject))))# %>%


dfh_avg$Orientation <- ordered(dfh_avg$Orientation, levels = c("Upright Face", "Inverted Face"))
dfh_avg$Condition <- ordered(dfh_avg$Condition, levels = c("Different","Same","Isolated"))

dfh$Orientation <- ordered(dfh$Orientation, levels = c("Upright Face", "Inverted Face"))
dfh$Condition <- ordered(dfh$Condition, levels = c("Different","Same","Isolated"))

mdata %>% ggplot(aes(x=iv, y=fit, group=Condition, color=Condition, fill=Condition))+
    geom_line(size=1.3, show.legend=TRUE, alpha=1)+
    geom_point(size=5, data=dfh_avg, aes(x=iv, y=mean), alpha=.5)+
    geom_errorbar(data=dfh_avg, inherit.aes=F, aes(x=iv, ymax=mean+se, ymin=mean-se), width=2.5, alpha=.3)+
    geom_ribbon(aes(ymin=right_lwr, ymax=right_upr), alpha=.25)+
    facet_wrap(~Orientation)+
    theme_classic() + 
    ggthemes::scale_color_colorblind()+
    ggthemes::scale_fill_colorblind()+
    scale_x_continuous(breaks=c(0,24,36,53,80))+
    scale_y_continuous(limits=c(0,1))+
    xlab("Dissimilarity (%)")+
    ylab("Proportion of 'Different' Responses")+
    theme(strip.background=element_blank(),
          strip.text=element_text(size=25),
          axis.title=element_text(size=20),
          axis.text=element_text(size=15))






```


# Thresholds - Residuals relationship

```{r Thresh-resid}

load("vectors_data.rda")
load("resids_data.rda")

vecsnew <- vecs4 %>% filter(task=="Inverted Face") %>%
  select(subject, cond, val) %>%
  pivot_wider(names_from=cond, values_from=val) %>%
  mutate(diff=Same-Different)

vecsresids <- tibble(vecsnew$subject, vecsnew$diff, b4$ldsresid) %>%
  rename(subject=1, diff=2, resid=3)

vecsresids %>% ggplot(aes(x=diff,y=resid))+
  geom_point()+
  geom_smooth(method=lm)


```
